// hosts/web/src/generated/HostMessagePayload.ts
// Generated by Passport. Do not modify.
// File generated at 2/28/2026, 11:05 AM
import type {HostMessagePermissionResponded} from "./HostMessagePermissionResponded"
import type {HostMessageSubscribeLogsAck} from "./HostMessageSubscribeLogsAck"
import type {HostMessageEventReceived} from "./HostMessageEventReceived"
import type {HostMessageHello} from "./HostMessageHello"
import type {HostMessageAuthorizationCreate} from "./HostMessageAuthorizationCreate"
import type {HostMessageSubscribeLogs} from "./HostMessageSubscribeLogs"
import type {HostMessageHelloAck} from "./HostMessageHelloAck"
import type {HostMessageAuthorizationCreated} from "./HostMessageAuthorizationCreated"
import type {HostMessageError} from "./HostMessageError"
export interface HostMessagePayload {
	hello: HostMessageHello | undefined
	hello_ack: HostMessageHelloAck | undefined
	subscribe_logs: HostMessageSubscribeLogs | undefined
	subscribe_logs_ack: HostMessageSubscribeLogsAck | undefined
	authorization_create: HostMessageAuthorizationCreate | undefined
	authorization_created: HostMessageAuthorizationCreated | undefined
	request_permission: HostMessagePermissionResponded | undefined
	permission_response: HostMessagePermissionResponded | undefined
	event_received: HostMessageEventReceived | undefined
	error: HostMessageError | undefined
}