// hosts/web/src/generated/HostMessageType.ts
// Generated by Passport. Do not modify.
// File generated at 2/28/2026, 3:02 PM

export type HostMessageType = "componentsList" | "getTasksList" | "permissionRequest" | "respondPermission" | "tasksList" | "getComponentsList" | "logEvent" | "authorizationsList" | "requestHistory" | "pendingPermissionList" | "getAuthorizationsList" | "getPendingPermissionList" | "getRequestHistory" | "createAuthorization"